+++ src/config/schema.ts
@@ -1,5 +1,7 @@
 import * as zod from "zod";
 
+import { StringMatcherOptionsSchema, StringMatcherSchema } from "../matching/stringMatcher";
+import { WordMatcherOptionsSchema, WordMatcherSchema } from "../matching/wordMatcher";
 import { DeepPartial } from "../utils/types";
 
 const pluginSchema = zod.object({
@@ -32,56 +34,6 @@ export const ApplyStudioLabelsEnum = zod.enum([
   "plugin:scene:create",
   "plugin:scene:custom",
 ]);
-
-const StringMatcherOptionsSchema = zod.object({
-  ignoreSingleNames: zod.boolean(),
-  stripString: zod.string(),
-});
-
-export type StringMatcherOptions = zod.TypeOf<typeof StringMatcherOptionsSchema>;
-
-const StringMatcherSchema = zod.object({
-  type: zod.enum(["legacy", "string"]),
-  options: StringMatcherOptionsSchema,
-});
-
-export type StringMatcherType = zod.TypeOf<typeof StringMatcherSchema>;
-
-const WordMatcherOptionsSchema = zod.object({
-  ignoreSingleNames: zod.boolean(),
-  ignoreDiacritics: zod.boolean(),
-  /**
-   * If word groups should be not used. Allows words to match across word groups.
-   * Example: allows "My WordGroup" to match against "My WordGroupExtra"
-   */
-  enableWordGroups: zod.boolean(),
-  /**
-   * If a group of words does not contain any group separators, if the word separators
-   * should be used to separate groups instead of words
-   */
-  wordSeparatorFallback: zod.boolean(),
-  /**
-   * If a camelCase word (PascalCase included) should create a word group
-   */
-  camelCaseWordGroups: zod.boolean(),
-  /**
-   * When inputs were matched on overlapping words, which one to return.
-   * Example: "My Studio", "Second My Studio" both overlap when matched against "second My Studio"
-   */
-  overlappingMatchPreference: zod.enum(["all", "longest", "shortest"]),
-  groupSeparators: zod.array(zod.string()),
-  wordSeparators: zod.array(zod.string()),
-  filepathSeparators: zod.array(zod.string()),
-});
-
-export type WordMatcherOptions = zod.TypeOf<typeof WordMatcherOptionsSchema>;
-
-const WordMatcherSchema = zod.object({
-  type: zod.literal("word"),
-  options: WordMatcherOptionsSchema,
-});
-
-export type WordMatcherType = zod.TypeOf<typeof WordMatcherSchema>;
 
 const logLevelType = zod.enum(["error", "warn", "info", "http", "verbose", "debug", "silly"]);
